rename calc_subblock_memory.py to calc_subblock_params_and_memory.py
